requirements.txt
ultralytics
opencv_python
Pillow
omegaconf

# product visual search
vertexai
chromadb
tqdm
numpy

# CLIP
torch
torchvision
git+https://github.com/openai/CLIP.git
ftfy
regex
tqdm
